remove public toilet only by creator and unpublished

+++ src/public-toilet/public-toilet.service.ts
@@ -1,5 +1,5 @@
 import { GeoJSONPointScalar } from "src/common/scalars/geojson/Point.scalar";
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { CreatePublicToiletInput } from './dto/create-public-toilet.input';
@@ -10,6 +10,8 @@ import {
 } from './entities/public-toilet.entity';
 import { uploadFileStream } from '../common/utils/upload';
 import { PublishState as PublicToiletState } from "src/common/enum/publish_state.enum";
+import { checkUserIsAuthor } from "src/common/utils/checkUserAuthentication";
+import { checkIfObjectIsPublished } from "src/common/utils/checkPublishedState";
 
 @Injectable()
 export class PublicToiletService {
@@ -181,7 +183,10 @@ export class PublicToiletService {
     throw new NotFoundException(`PublicToilet with id ${id} not found`);
   }
 
-  async remove(id: number) {
+  async remove(
+    id: number,
+    user: number,
+  ) {
     const publicToilet: PublicToilet =
       await this.publicToiletRepository.findOne({
         where: { id: id },
@@ -189,6 +194,11 @@ export class PublicToiletService {
       });
 
     if (publicToilet) {
+      const published_state = checkIfObjectIsPublished(publicToilet.state);
+      if (published_state) {
+        return new ForbiddenException(`PublicToilet with id ${id} already published.`)
+      }
+      checkUserIsAuthor(user, publicToilet.createdBy);
       const deleteImage = publicToilet.images.map(async (image) => {
         return await this.publicToiletImageRepository.delete(image.id);
       });
